movierecomm-app/requirements.txt
Flask
gunicorn
numpy
pandas
random
sklearn.metrics.pairwise
sklearn.impute
sklearn.decomposition
pickle
